packetizer: mpeg4audio: remove unnecessary cast
Both variables are size_t since 48452d5a7cd86f7ba8a77f00525196f94b1e5622
and 03bba25b5d327f7f18c585eb49d57fe32f90d012.

+++ modules/packetizer/mpeg4audio.c
@@ -452,7 +452,7 @@ static int LOASParse(decoder_t *p_dec, uint8_t *p_buffer, int i_buffer)
 
         if (p_sys->i_channels && p_sys->i_rate && p_sys->i_frame_length > 0)
         {
-            if((size_t)p_dec->fmt_out.i_extra != st->i_extra ||
+            if(p_dec->fmt_out.i_extra != st->i_extra ||
                (p_dec->fmt_out.i_extra > 0 &&
                 memcmp(p_dec->fmt_out.p_extra, st->extra, st->i_extra)) )
             {
